Struct files propertly
