+++ src/BundlerMinifier.Core/project.json
@@ -30,7 +30,8 @@
         "Microsoft.NETCore.App": {
           "type": "platform",
           "version": "1.1.0"
-        }
+        },
+        "System.Reflection.TypeExtensions": "4.1.0"
       }
     },
     "net452": {}
